Fix really stupid bug affecting recursion
Basically, the fact that parsed Scheem is passed to the interpreter as an array wasn't on my mind... So, I never worried about calling .shift(). But objects are all shallow copied in JS, so that affected everyone. So during a second recursion pass, lambda would try to access args[1], which didn't work well because args (aka expr) had been shifted. End result: trying to .apply() stuff such as numbers.